split stores to files, deleted unused utils
Moved all stores to individual files.
Deleted utils, superseded by stores.
